transdimensional summoning
(but STILL not able to keep your request when you go through a portal because Neo Doesn't Love Me. but whatever it's naturally incompatible of course you can't maintain a summon target if it gets summoned elsewhere, there's no way to keep track of these kinds of things. it's fine this is intended behavior. I WANTED THIS. i'll maybe figure it out later or maybe beg for salvation from someone else but it's not that big a deal, seems unlikely that you'd go through a portal while having a request and want to keep that request