fabrication
most things are identical and my neo testing worlds were nearly perfectly intact when loading them with fabric, but there are a few things different.

* moved megumin's hat to its own little section that only loads if fbombs is around. it belongs to fbombs it deserves this
* fixed the singed component being needlessly complex, which breaks old singed tags
* removed true naming functionality from name tags (as well as their dispenser behavior) because that was a little silly. for now, you can just craft a name tag into a true name
* added hats for slimes and magma cubes. very important

i think that's it? tried to not change too much for the big fabrication commit but whatevr. see ya in a bit when i care about connector, neo!